+++ app/controllers/patients/pipelines_controller.rb
@@ -3,6 +3,7 @@
 module Patients
   class PipelinesController < ApplicationController
     before_action :load_patient
+    before_action :load_patient_details, only: [:show, :new, :edit]
     before_action :find_and_authorize, only: [:show, :edit, :update, :destroy]
 
     def index
@@ -65,6 +66,10 @@ module Patients
 
     def load_patient
       @patient = Patient.find_by!(case_number: params[:patient_id])
+    end
+
+    def load_patient_details
+      @details = Patients::Details.new(@patient.case_number, current_user).call
     end
 
     def find_and_authorize