+++ spec/requests/patients_spec.rb
@@ -8,7 +8,7 @@ describe 'Patients controller' do
 
   context 'with no user signed in' do
     describe 'GET /patients' do
-      it 'is redirects to signin url' do
+      it 'redirects to sign-in url' do
         get '/patients'
         expect(response).to redirect_to new_user_session_path
       end
@@ -58,6 +58,46 @@ describe 'Patients controller' do
           post '/patients/', params: { patient: { case_number: '5555' } }
         end.to change { Patient.count }.by(1)
         expect(response).to redirect_to Patient.first
+      end
+    end
+
+    describe 'external data sets service with patient details' do
+      it 'is called and returns empty result set' do
+        expect_any_instance_of(Patients::Details).to receive(:call).and_return(
+          status: :error,
+          message: 'reason'
+        )
+
+        get "/patients/#{patient.case_number}"
+
+        expect(response.body).to include(I18n.t('patients.show.no_details', details: 'reason'))
+      end
+
+      it 'is called and returns valid results' do
+        expect_any_instance_of(Patients::Details).to receive(:call).and_return(
+          status: :ok,
+          payload: [
+            { name: 'gender', value: 'Male', type: 'real', style: 'default' },
+            { name: 'birth_year', value: 1970, type: 'real', style: 'default' },
+            { name: 'age', value: 47, type: 'real', style: 'default' },
+            { name: 'current_age', value: 50, type: 'computed', style: 'success' },
+            { name: 'height', value: 170, type: 'real', style: 'default' },
+            { name: 'weight', value: 45, type: 'real', style: 'real' },
+            { name: 'bpprs', value: 20, type: 'inferred', style: 'warning' },
+            { name: 'bpprd', value: 30, type: 'inferred', style: 'warning' }
+          ]
+        )
+
+        get "/patients/#{patient.case_number}"
+
+        expect(response.body).to include("#{I18n.t('patients.show.gender')}: Male")
+        expect(response.body).to include("#{I18n.t('patients.show.birth_year')}: 1970")
+        expect(response.body).to include("#{I18n.t('patients.show.age')}: 47")
+        expect(response.body).to include("#{I18n.t('patients.show.current_age')}: 50")
+        expect(response.body).to include("#{I18n.t('patients.show.height')}: 170")
+        expect(response.body).to include("#{I18n.t('patients.show.weight')}: 45")
+        expect(response.body).to include("#{I18n.t('patients.show.bpprs')}: 20")
+        expect(response.body).to include("#{I18n.t('patients.show.bpprd')}: 30")
       end
     end
   end